pyproject.toml
[project]
name = "count-code-lines"
version = "0.0.4"
description = "A small package and CLI for counting code lines"
readme = "README.md"
requires-python = "==3.12"
dependencies = [
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cryptography==43.0.1",
    "gitdb==4.0.11",
    "github3-py==4.0.1",
    "gitpython==3.1.43",
    "idna==3.10",
    "loguru==0.7.2",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "millify==0.1.1",
    "pip==23.2.1",
    "plotext==5.3.2",
    "py-markdown-table==1.1.0",
    "pycparser==2.22",
    "pygments==2.18.0",
    "pygount==1.8.0",
    "pyjwt==2.9.0",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.3",
    "rich==13.9.1",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smmap==5.0.1",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
]

[project.scripts]
count-code-lines = "count_code_lines:cli_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
